fix: align SQLite node with database nodes
