+++ backend/app/nodes/databases/__init__.py
@@ -1,5 +1,6 @@
 """Database integration nodes."""
 
 from .postgres_node import PostgresNode
+from .sqlite_node import SQLiteNode
 
-__all__ = ["PostgresNode"]
+__all__ = ["PostgresNode", "SQLiteNode"]
